move categorical risk observer to its own file. Add config blocks.
