+++ src/vivarium_conic_sam_comparison/components/metrics/__init__.py
@@ -1,1 +1,2 @@
-from .disability import WHZDisabilityObserver, CatStratRiskObserver
+from .disability import WHZDisabilityObserver
+from .risk import CatStratRiskObserver
